backend/requirements.txt
fastapi
uvicorn[standard]
python-multipart
pydantic
python-dotenv
supabase
google-genai
mistralai
psycopg[binary]
